requirements.txt
flask==1.1.2
pytesseract==0.3.7
Pillow==8.1.2
